# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SK Physics API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Note {
  id: number;
  title: string;
  description: string;
  fileUrl: string;
  fileName: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNoteBody {
  title: string;
  description: string;
  fileUrl: string;
  fileName: string;
  adminToken: string;
}

export interface UpdateNoteBody {
  title: string;
  description: string;
  adminToken: string;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResponse {
  token: string;
  success: boolean;
}

export interface UploadUrlBody {
  fileName: string;
  adminToken: string;
}

export interface UploadUrlResponse {
  uploadUrl: string;
  fileUrl: string;
}