# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SK Physics API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all available physics notes
 * @summary List all notes
 */
export const ListNotesResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  fileUrl: zod.string(),
  fileName: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListNotesResponse = zod.array(ListNotesResponseItem);

/**
 * Upload a new physics note (admin only)
 * @summary Create a new note
 */
export const CreateNoteBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  fileUrl: zod.string(),
  fileName: zod.string(),
  adminToken: zod.string(),
});

/**
 * @summary Get a specific note
 */
export const GetNoteParams = zod.object({
  id: zod.coerce.number(),
});

export const GetNoteResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  fileUrl: zod.string(),
  fileName: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a note
 */
export const UpdateNoteParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateNoteBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  adminToken: zod.string(),
});

export const UpdateNoteResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  fileUrl: zod.string(),
  fileName: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a note
 */
export const DeleteNoteParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  token: zod.string(),
  success: zod.boolean(),
});

/**
 * @summary Get a signed upload URL for PDF
 */
export const GetUploadUrlBody = zod.object({
  fileName: zod.string(),
  adminToken: zod.string(),
});

export const GetUploadUrlResponse = zod.object({
  uploadUrl: zod.string(),
  fileUrl: zod.string(),
});