# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SK Physics API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Note {
  id: number;
  title: string;
  description: string;
  fileUrl: string;
  fileName: string;
  createdAt: string;
  updatedAt: string;
}