# lib/api-zod/src/generated/types/uploadUrlBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SK Physics API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UploadUrlBody {
  fileName: string;
  adminToken: string;
}